requirements.txt
fastapi
pydantic==1.*
python-multipart
uvicorn
tortoise-orm